feat: support migrating tags in `kopia snapshot migrate`
This PR adds a new command line option to `kopia snapshot migrate` named
`--tags`. This new flag allows the user to specify a list of tags to be
migrated at the same sime as the snapshots.

If no tag is specified no tag will be migrated, preserving the current
behaviour.

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>